style: :truck: Moves all files one folder up to the main folder and move the services into the control package
