+++ .gitignore
@@ -1,24 +1,43 @@
-# Compiled class file
-*.class
+#Maven
+target/
+pom.xml.tag
+pom.xml.releaseBackup
+pom.xml.versionsBackup
+release.properties
+.flattened-pom.xml
 
-# Log file
-*.log
+# Eclipse
+.project
+.classpath
+.settings/
+bin/
 
-# BlueJ files
-*.ctxt
+# IntelliJ
+.idea
+*.ipr
+*.iml
+*.iws
 
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
+# NetBeans
+nb-configuration.xml
 
-# Package Files #
-*.jar
-*.war
-*.nar
-*.ear
-*.zip
-*.tar.gz
-*.rar
+# Visual Studio Code
+.vscode
+.factorypath
 
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
-replay_pid*
+# OSX
+.DS_Store
+
+# Vim
+*.swp
+*.swo
+
+# patch
+*.orig
+*.rej
+
+# Local environment
+.env
+
+# Plugin directory
+/.quarkus/cli/plugins/
