+++ src/main/docker/Dockerfile.legacy-jar
