+++ src/main/docker/Dockerfile.native-micro
