+++ src/main/java/hftm/blog/boundary/BlogResource.java
