+++ src/main/java/hftm/blog/entity/Comment.java
@@ -1,0 +1,15 @@
+/* package hftm.blog.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+
+@Entity
+public class Comment {
+    
+    @ManyToOne
+    @JoinColumn()
+    private Blog blog;
+
+}
+ */
