+++ src/main/resources/META-INF/resources/index.html
