+++ src/test/java/hftm/blog/AuthorBlogTest.java
@@ -4,8 +4,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import hftm.blog.boundary.AuthorService;
-import hftm.blog.boundary.BlogService;
+import hftm.blog.control.AuthorService;
+import hftm.blog.control.BlogService;
 import hftm.blog.entity.Author;
 import hftm.blog.entity.Blog;
 import io.quarkus.test.junit.QuarkusTest;
