+++ src/test/java/hftm/blog/BlogServiceTest.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import hftm.blog.boundary.BlogService;
+import hftm.blog.control.BlogService;
 import hftm.blog.entity.Blog;
 import io.quarkus.test.junit.QuarkusTest;
 import jakarta.inject.Inject;
